+++ src/msg/msg_comm.cpp
@@ -84,7 +84,7 @@ int MSG_comm_test(msg_comm_t comm)
  */
 int MSG_comm_testany(const_xbt_dynar_t comms)
 {
-  int finished_index = -1;
+  ssize_t finished_index = -1;
 
   /* Create the equivalent array with SIMIX objects: */
   std::vector<simgrid::kernel::activity::CommImpl*> s_comms;
@@ -120,7 +120,7 @@ int MSG_comm_testany(const_xbt_dynar_t comms)
     }
   }
 
-  return finished_index;
+  return static_cast<int>(finished_index);
 }
 
 /** @brief Destroys the provided communication. */
